Moved test_collect_nodeid() to test_plugin.py
