+++ tests/test_plugin.py
@@ -4,6 +4,7 @@ from pathlib import Path
 import sys
 
 import pytest_phmdoctest
+import pytest_phmdoctest.collectors
 import pytest
 
 
@@ -520,6 +521,24 @@ def test_file_has_no_examples(pytester):
     rr.stdout.fnmatch_lines(
         [
             "*collected 0 items*",
+        ],
+        consecutive=True,
+    )
+
+
+def test_collect_nodeid(pytester, file_creator):
+    """Collect a generated test file by nodeid."""
+
+    file_creator.populate_all(pytester_object=pytester)
+    rr = pytester.runpytest("--phmdoctest-docmod", "-v", "-k", "README.py")
+    if pytest_phmdoctest.collectors.PYTEST_GE_7:
+        rr.assert_outcomes(passed=2, deselected=8)
+    else:
+        rr.assert_outcomes(passed=2)
+    rr.stdout.fnmatch_lines(
+        [
+            "*::README.py::README.session_00001_line_24 PASSED*",
+            "*::README.py::test_code_10_output_17 PASSED*",
         ],
         consecutive=True,
     )
